chore(rxPage): multiple fix
